feat!: rewrite codebase
This commit is a complete rewrite of the codebase for the sake of making it more minimal, with better separation of concerns and less code duplication. A new command-line tool for interacting with the daemon was created to make `blctl` more user-friendly. D-Bus was dropped in favor of Unix domain sockets for IPC between the client (`blctl`, the command-line tool) and the daemon (`blctld`). Additionally, `blctl` is now packaged for NixOS using a nix flake.